extract SearchResultsPanel from MallSearchFrame to make it testable
